requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
reportlab==4.4.1
google-generativeai==0.5.2
